ULTIMA4: Move DungeonView into views/
