+++ engines/ultima/ultima4/game/game.cpp
@@ -64,9 +64,9 @@
 #include "ultima/ultima4/map/shrine.h"
 #include "ultima/ultima4/map/tilemap.h"
 #include "ultima/ultima4/map/tileset.h"
-#include "ultima/ultima4/map/dungeonview.h"
 #include "ultima/ultima4/sound/music.h"
 #include "ultima/ultima4/sound/sound.h"
+#include "ultima/ultima4/views/dungeonview.h"
 #include "ultima/ultima4/meta_engine.h"
 #include "common/savefile.h"
 #include "common/system.h"
