+++ engines/ultima/ultima4/gfx/screen.cpp
@@ -33,10 +33,10 @@
 #include "ultima/ultima4/gfx/imagemgr.h"
 #include "ultima/ultima4/gfx/screen.h"
 #include "ultima/ultima4/gfx/textcolor.h"
-#include "ultima/ultima4/map/dungeonview.h"
 #include "ultima/ultima4/map/location.h"
 #include "ultima/ultima4/map/tileanim.h"
 #include "ultima/ultima4/map/tileset.h"
+#include "ultima/ultima4/views/dungeonview.h"
 #include "ultima/ultima4/views/tileview.h"
 #include "ultima/ultima4/map/annotation.h"
 #include "ultima/ultima4/ultima4.h"
